DrawingTask: extract DrawingTool into own subcomponent
